+++ ChangeLog.md
@@ -1,3 +1,10 @@
 # Changelog for Loopbreaker
 
+## 0.1.1.0
+- follow `INLINE` pragmas when adding loopbreakers
+- support local declarations (`let` and `where` bindings)
+
+## 0.1.0.0
+- initial release
+
 ## Unreleased changes